backend/core/requirements.txt
Django==5.2
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
PyJWT==2.9.0
psycopg2==2.9.10

# Data Science and ML libraries
tensorflow==2.19.0
scikit-learn==1.3.2
numpy==1.26.2
pandas==2.1.3
matplotlib==3.8.2

dotenv==0.9.9
requests==2.32.3
google-generativeai==0.7.2